total_0520

+++ back/app/src/main/java/com/sloway/app/place/dto/request/station/StationReqDto.java
@@ -38,7 +38,7 @@ public class StationReqDto {
     @Getter
     @Setter
     public static class AmenityDto {
-        private int amenityNo;
+        private Long amenityNo;
     }
 
     @Getter
